product_api/requirements.txt
fastapi==0.110.3
uvicorn==0.20.0
pydantic==2.7.1
python-dotenv==1.0.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1